+++ outputs.tf
@@ -10,22 +10,22 @@ output "appgw_name" {
 
 output "appgw_subnet_id" {
   description = "The ID of the subnet where the Application Gateway is attached."
-  value       = var.create_subnet == true ? join("", module.azure-network-subnet.subnet_ids) : var.subnet_id
+  value       = var.create_subnet == true ? join("", module.azure_network_subnet.subnet_ids) : var.subnet_id
 }
 
 output "appgw_subnet_name" {
   description = "The name of the subnet where the Application Gateway is attached."
-  value       = var.create_subnet == true ? join("", module.azure-network-subnet.subnet_names) : null
+  value       = var.create_subnet == true ? join("", module.azure_network_subnet.subnet_names) : null
 }
 
 output "appgw_nsg_id" {
   description = "The ID of the network security group from the subnet where the Application Gateway is attached."
-  value       = module.azure-network-security-group.network_security_group_id
+  value       = module.azure_network_security_group.network_security_group_id
 }
 
 output "appgw_nsg_name" {
   description = "The name of the network security group from the subnet where the Application Gateway is attached."
-  value       = module.azure-network-security-group.network_security_group_name
+  value       = module.azure_network_security_group.network_security_group_name
 }
 
 output "appgw_public_ip_address" {