+++ modules/app-gateway-v1/outputs.tf
@@ -5,5 +5,5 @@ output "app_gateway_id" {
 
 output "network_security_group_id" {
   description = "Network Security Group ID of the subnet where is Application Gateway"
-  value       = module.network-security-group.network_security_group_id
+  value       = module.network_security_group.network_security_group_id
 }