+++ variables.tf
@@ -82,12 +82,6 @@ variable "ip_allocation_method" {
 
 # Application gateway inputs
 
-variable "appgw_name" {
-  description = "Application Gateway name."
-  type        = string
-  default     = ""
-}
-
 variable "sku_capacity" {
   description = "The Capacity of the SKU to use for this Application Gateway - which must be between 1 and 10, optional if autoscale_configuration is set"
   type        = number
@@ -104,24 +98,6 @@ variable "zones" {
   description = "A collection of availability zones to spread the Application Gateway over. This option is only supported for v2 SKUs"
   type        = list(string)
   default     = ["1", "2", "3"]
-}
-
-variable "frontend_ip_configuration_name" {
-  description = "The Name of the Frontend IP Configuration used for this HTTP Listener."
-  type        = string
-  default     = ""
-}
-
-variable "frontend_priv_ip_configuration_name" {
-  description = "The Name of the private Frontend IP Configuration used for this HTTP Listener."
-  type        = string
-  default     = ""
-}
-
-variable "gateway_ip_configuration_name" {
-  description = "The Name of the Application Gateway IP Configuration."
-  type        = string
-  default     = ""
 }
 
 variable "frontend_port_settings" {
@@ -326,12 +302,6 @@ variable "route_table_ids" {
   default     = {}
 }
 
-variable "custom_subnet_name" {
-  description = "Custom name for the subnet."
-  type        = string
-  default     = ""
-}
-
 variable "subnet_cidr" {
   description = "Subnet CIDR to create."
   type        = string
@@ -354,24 +324,6 @@ variable "create_nsg_healthprobe_rule" {
   description = "Boolean to create the network security group rule for the health probes."
   type        = bool
   default     = true
-}
-
-variable "custom_nsg_name" {
-  description = "Custom name for the network security group."
-  type        = string
-  default     = null
-}
-
-variable "custom_nsr_https_name" {
-  description = "Custom name for the network security rule for HTTPS protocol."
-  type        = string
-  default     = null
-}
-
-variable "custom_nsr_healthcheck_name" {
-  description = "Custom name for the network security rule for internal health check of Application Gateway."
-  type        = string
-  default     = null
 }
 
 variable "nsr_https_source_address_prefix" {