+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+Breaking
+  * AZ-546: Clean module, remove unused variables, needs a `terraform state mv` for renamed modules
+
+Fixed
+  * AZ-530: Fix provider required version
+
 # v4.8.0 - 2021-06-25
 
 Added