+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Breaking
   * AZ-546: Clean module, remove unused variables, needs a `terraform state mv` for renamed modules
+  * AZ-521: Revamp variables names, module cleanup
 
 Fixed
   * AZ-530: Fix provider required version